+++ rust/src/years/y2024/day02.rs
@@ -83,7 +83,7 @@ fn part02(file_contents: &str) -> u32 {
 }
 
 
-pub fn get_day_results(file_contents: &str) -> (u32, u32) {
+pub fn get_day_results(file_contents: &str) -> (u64, u64) {
     /*
         Return this day's results as a tuple
      */
